Move IntList to yggdrasil
